pyproject.toml
[tool.poetry]
name = "parque-marino-backend"
version = "1.0.0"
description = "Backend system for Parque Marino - A marine park management system"
authors = ["Your Name <your.email@example.com>"]
license = "Proprietary"

[tool.poetry.dependencies]
python = "^3.10"
Django = "5.2.3"
djangorestframework = "3.16.0"
djangorestframework-simplejwt = "5.5.0"
psycopg2-binary = "2.9.10"
mysqlclient = "2.2.7"
boto3 = "1.40.16"
django-storages = "1.14.6"
django-redis = "6.0.0"
stripe = "10.10.0"
twilio = "8.8.0"
drf-spectacular = "0.28.0"
drf-yasg = "1.21.10"
python-decouple = "3.8"
python-dotenv = "0.19.2"
gunicorn = "23.0.0"
whitenoise = "6.9.0"
django-reversion = "5.1.0"
django-cors-headers = "4.6.0"
dj-database-url = "2.3.0"
Pillow = "11.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-django = "^4.9.0"
black = "^24.8.0"
flake8 = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
python_files = "tests.py test_*.py *_tests.py"
DJANGO_SETTINGS_MODULE = "config.settings.development"
testpaths = [
    "tests",
    "apps",
]